# packages/backend/swagger.json
{
	"openapi": "3.0.0",
	"info": {
		"version": "0.1.0",
		"title": "WAFRN Backend",
		"description": "API routes for wafrn social network"
	},
	"paths": {
...
		"/api/admin/server-list": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/admin/server-update": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/admin/userBlockList": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/admin/reportCount": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/admin/reportList": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/admin/closeReport": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/banUser": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/ignoreReport": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/getBannedUsers": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/admin/unbanUser": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/getPendingApprovalUsers": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/admin/activateUser": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/admin/notActivateAndSendEmail": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"id": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/blockUserServer": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/unblockUserServer": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				},
				"requestBody": {
					"content": {
						"application/json": {
							"schema": {
								"type": "object",
								"properties": {
									"userId": {
										"example": "any"
									}
								}
							}
						}
					}
				}
			}
		},
		"/api/myServerBlocks": {
			"get": {
				"description": "",
				"parameters": [
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
		"/api/unblockServer": {
			"post": {
				"description": "",
				"parameters": [
					{
						"name": "id",
						"in": "query",
						"schema": {
							"type": "string"
						}
					},
					{
						"name": "authorization",
						"in": "header",
						"schema": {
							"type": "string"
						}
					}
				],
				"responses": {
					"200": {
						"description": "OK"
					},
					"401": {
						"description": "Unauthorized"
					}
				}
			}
		},
...
	},
	"components": {
		"securitySchemes": {
			"bearerAuth": {
				"type": "http",
				"scheme": "bearer",
				"bearerFormat": "JWT"
			}
		}
	}
}